Removes outgoing and view_actions
